requirements.txt
Django==4.2.16
django-environ==0.12.0
gunicorn==23.0.0
langchain==0.3.15
langchain-community==0.3.15
langchain-openai==0.3.2
langsmith==0.3.2
llama-index==0.12.14
llama-index-vector-stores-postgres==0.4.2
psycopg2-binary==2.9.10
python-decouple==3.8
